+++ modules/task/src/main/java/com/linln/modules/task/repository/PriceRepository.java
@@ -12,11 +12,11 @@ import java.util.List;
 public interface PriceRepository extends BaseRepository<Price,Long>, JpaSpecificationExecutor<Price> {
 
     @Query(value = "select * from price where price_start<=?1 and price_end >?1  and price_type=?2 and system_name='B'",nativeQuery = true)
-  List<Price> getMerchantPriceByPrice(int price, String priceType);
+  List<Price> getMerchantPriceByPrice(BigDecimal price, String priceType);
 
 
     @Query(value = "select * from price where price_start<=?1 and price_end >?1  and price_type=?2 and system_name='C'",nativeQuery = true)
-    List<Price> getCustomerPriceByPrice(int price, String priceType);
+    List<Price> getCustomerPriceByPrice(BigDecimal price, String priceType);
 
     @Modifying
     @Transactional
