+++ modules/task/src/main/java/com/linln/modules/task/service/PriceService.java
@@ -4,6 +4,7 @@ import com.linln.modules.task.domain.Price;
 import org.springframework.data.domain.Example;
 import org.springframework.data.domain.Page;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface PriceService {
@@ -26,9 +27,9 @@ public interface PriceService {
 
     Price getMerchantPriceById(Long id);
 
-    List<Price> getMerchantPriceByPrice(int price, String priceType);
+    List<Price> getMerchantPriceByPrice(BigDecimal price, String priceType);
 
-    List<Price> getCustomerPriceByPrice(int price, String priceType);
+    List<Price> getCustomerPriceByPrice(BigDecimal price, String priceType);
 
 
     int update(Price price);
